+++ Back/Latte/src/main/java/com/latte/admin/web/dto/order/OrderedUuidResponseDto.java
@@ -11,7 +11,7 @@ public class OrderedUuidResponseDto {
 
     private int ostatus;  // 주문상태: -1=취소, 0=대기, 1=사장님확인, 2=완료
     private String ocontent;
-    private Long oprice;
+    private int oprice;
     private Long ccid;
 
     @Builder
